feat(cli): Keep looking in parent folder for webpack.client.config.js if not found.
This feature is to fix flareact command in monorepo project, where all dependencies are aggregated in root folder node_modules. So we need to find the config file in parent folder.

I use while loop that will stop when the config is found and keep traversing parent folder if not yet found.